+++ src/Config/Config.php
@@ -2,11 +2,14 @@
 
 namespace TarfinLabs\LaravelConfig\Config;
 
+use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\Model;
 use TarfinLabs\LaravelConfig\Casts\ConfigValueCast;
+use TarfinLabs\LaravelConfig\Database\Factories\ConfigFactory;
 
 class Config extends Model
 {
+    use HasFactory;
     protected $table = 'config';
 
     protected $guarded = [];
@@ -22,5 +25,10 @@ class Config extends Model
     {
         $this->setTable(config('laravel-config.table'));
         parent::__construct($attributes);
+    }
+
+    protected static function newFactory(): ConfigFactory
+    {
+        return ConfigFactory::new();
     }
 }