+++ src/LaravelConfigServiceProvider.php
@@ -20,10 +20,6 @@ class LaravelConfigServiceProvider extends ServiceProvider
             $this->publishes([
                 __DIR__.'/../config/config.php' => config_path('laravel-config.php'),
             ], 'laravel-config');
-
-            $this->publishes([
-                __DIR__.'/../database/factories/ConfigFactory.php' => database_path('factories/ConfigFactory.php'),
-            ], 'laravel-config-factories');
         }
     }