+++ tests/TestCase.php
@@ -11,8 +11,6 @@ class TestCase extends \Orchestra\Testbench\TestCase
     {
         parent::setUp();
 
-        $this->withFactories(__DIR__.'/../database/factories');
-
         Schema::dropAllTables();
 
         $this->artisan('migrate', [